Move exported types out of route.ts file
