+++ src/app/api/users/route.ts
@@ -1,17 +1,6 @@
 import { clerkClient } from "@clerk/nextjs/server";
-import { z } from "zod";
 
-export const getUsersSchema = z.object({
-  userId: z.array(z.string()).optional(),
-  username: z.array(z.string()).optional(),
-  emailAddress: z.array(z.string()).optional(),
-});
-
-export type User = {
-  id: string;
-  name: string;
-  imageUrl: string;
-};
+import { getUsersSchema } from "./types";
 
 export async function GET(req: Request) {
   try {
