+++ src/hooks/useUsers.ts
@@ -1,7 +1,7 @@
 import { useQuery } from "@tanstack/react-query";
 import { type z } from "zod";
 
-import { type User, type getUsersSchema } from "~/app/api/users/route";
+import { type User, type getUsersSchema } from "~/app/api/users/types";
 
 export function getUsers(params: z.infer<typeof getUsersSchema>) {
   const queryString = new URLSearchParams();
